pyproject.toml
[project]
name = "aivalink"
version = "0.1.0"
description = "AivaLink - AI Virtual Assistant Link Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=44.0.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "email-validator>=2.0.0",
    "openai>=1.0.0",
    "edge-tts>=6.1.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
gpu = [
    "faster-whisper>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "testcontainers[postgres]>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
strict = false

[tool.coverage.run]
concurrency = ["greenlet", "thread"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
